requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
asyncio-nats-client==0.11.4
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
coverage==5.5
dnspython==2.1.0
ed25519==1.5
email-validator==1.1.2
fastapi==0.63.0
flake8==3.9.1
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
httptools==0.1.1
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.3
nkeys==0.1.0
orjson==3.5.0
packaging==20.9
pluggy==0.13.1
promise==2.3
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.1
pyflakes==2.3.1
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.2.3
pytest-cov==2.11.1
pytest-mock==3.6.0
python-dotenv==0.15.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.25.1
Rx==1.6.1
six==1.15.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
ujson==3.2.0
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
